Improve default zooming
When we have bunch of users, we want the default zoom to allow to see
all of them, but when we do not have many users, the zoom is generaly
not appropriate.

Add a configuration setting to allow to choose when fitting the world is
preferred to fitting to the locations of all users.