Always run action on the gh-pages branch
This branch has the actual data. It should be the default branches of
forked repositories, but the main repo will use the main branch for that
purpose.  With this, we should not have data pushed to the wrong
location.

While here remove fetch-depth as 0 means fetch everything and the
default is more sensible.