Make map wrapping smoother
Enable `worldCopyJump` which allow markers to jump from their current
map to the "previous" or "next" one when scrolling.  This allow the map
to wrap, but when this happen, markers jump from the old position to the
new one.  To avoid this, add each marker twice: one time on the
"current" map, and one more time on the "previous" or the "next" one
depending on the side of the earth it on.

This allows us to have markers at the same time on the left and on the
right of the antimeridian, and when scrolling over this line the side
effects of the markers jumps are not too much impactful.