Swith to relative URL
Absolute URL are not ideal because when forking the repository, the
first deployment may use the domain name of the upstream project, and
while the served files are the correct ones, markers will be fetched
from the upstream project, and the map will display unexpected data.

By switching to relative URL, we download an empty list of positions,
which result in no markers on the map which is expected.

Also when running in development mode, we can use either 127.0.0.1, ::1,
or localhost (or any other name) without the need to configure
Cross-Origin Resource Sharing (CORS).  This was due to `jekyll serve`
using "localhost" as url, while the command output that the webiste is
available on 127.0.0.1 (so with a different name).